Avoid compilation on jazzy or higher (#301)
On jazzy the released openvdb version is already 10.0.1, so we don't
have to compile the vendor package.

Compilation on the build farm was already quite flakey because the
package is quite heavy, so I'd be best to avoid this. I don't know if
openvdb is also released for ARM or any other ubuntu platforms, but
we'll have to see on the buildfarm.